Update break-infinity-upgrades.js
Hello Vsauce, Michael here. I am still porting to Break Eternity, but what is eternity? Well imagine you have one replicanti in front of you. If that replicanti started replicating, as it always does, it would get to Eternity in exactly 5 hours. The act of breaking eternity, which is the equialent of the replicanti replicating for more than 5 hours, is a very new and dangerous technology that could completely destroy the universe, leaving only the expanse and the Elementals if it is not used carefully. If we did that, I bet the Elementals would be really annoyed! hahaha!